requirements.txt
huggingface_hub>=0.25.0
python-dotenv>=1.0.1 
huggingface-hub
python-dotenv 
requests 